+++ msm/src/serialization/witness.rs
@@ -34,8 +34,8 @@ impl<const N: usize, Fp: Field> InterpreterEnv for Env<N, Fp> {
     fn bitmask(
         &mut self,
         x: &Self::Variable,
-        highest_bit: u128,
-        lowest_bit: u128,
+        highest_bit: u32,
+        lowest_bit: u32,
         position: Self::Position,
     ) -> Self::Variable {
         let x: u128 = *x;
